Added development distance/time analysis, updated documentation with guide for pyH2A and small bug fixes

+++ src/pyH2A/Plugins/Photovoltaic_Plugin.py
@@ -6,7 +6,7 @@ class Photovoltaic_Plugin:
 
 	Parameters
 	----------
-	Financial Input Values > construction time > Value : float
+	Financial Input Values > construction time > Value : int
 		Construction time of hydrogen production plant in years.
 	Irradiation Used > Data > Value : str or ndarray
 		Hourly power ratio data for electricity production calculation. Either a
